since vul_file::size returns unsigned long, inline vul_file_size, should also return unsigned long

+++ core/vul/vul_file.h
@@ -104,8 +104,8 @@ inline bool vul_file_exists(vcl_string  f) { return vul_file::exists(f); }
 inline bool vul_file_is_directory(char const *f) { return vul_file::is_directory(f); }
 inline bool vul_file_is_directory(vcl_string  f) { return vul_file::is_directory(f); }
 
-inline int vul_file_size(char const *f) { return vul_file::size(f); }
-inline int vul_file_size(vcl_string  f) { return vul_file::size(f); }
+inline unsigned long vul_file_size(char const *f) { return vul_file::size(f); }
+inline unsigned long vul_file_size(vcl_string  f) { return vul_file::size(f); }
 
 inline vcl_string vul_file_extension(char const *f) { return vul_file::extension(f); }
 inline vcl_string vul_file_extension(vcl_string  f) { return vul_file_extension(f.c_str()); }
